[fe] refactor : 컴포넌트를 페이지별로 관리하기 위해 리팩토링 - 공통으로 사용되는 컴포넌트는 common으로, - 페이지 안에서만 사용되는 컴포넌트는 components/pageName 안에서 관리한다
